+++ frontend/src/pages/Notifications/actions.js
@@ -20,11 +20,11 @@ export const FETCH_ALL_NOTIFICATIONS_SUCCESS = "FETCH_ALL_NOTIFICATIONS_SUCCESS"
 export const LIVE_UPDATE_NOTIFICATIONS = "LIVE_UPDATE_NOTIFICATIONS";
 export const LIVE_UPDATE_NOTIFICATIONS_SUCCESS = "LIVE_UPDATE_NOTIFICATIONS_SUCCESS";
 
-export const MARK_MULTIPLE_NOTIFICATION_AS_READ = "MARK_MULTIPLE_NOTIFICATION_AS_READ";
-export const MARK_MULTIPLE_NOTIFICATION_AS_READ_SUCCESS = "MARK_MULTIPLE_NOTIFICATION_AS_READ_SUCCESS";
+export const MARK_MULTIPLE_NOTIFICATIONS_AS_READ = "MARK_MULTIPLE_NOTIFICATIONS_AS_READ";
+export const MARK_MULTIPLE_NOTIFICATIONS_AS_READ_SUCCESS = "MARK_MULTIPLE_NOTIFICATIONS_AS_READ_SUCCESS";
 
-export const FETCH_NOTIFICATION_COUNTS = "FETCH_NOTIFICATION_COUNTS";
-export const FETCH_NOTIFICATION_COUNTS_SUCCESS = "FETCH_NOTIFICATION_COUNTS_SUCCESS";
+export const FETCH_NOTIFICATION_COUNT = "FETCH_NOTIFICATION_COUNT";
+export const FETCH_NOTIFICATION_COUNT_SUCCESS = "FETCH_NOTIFICATION_COUNT_SUCCESS";
 
 export const SET_NOTIFICATIONS_PER_PAGE = "SET_NOTIFICATIONS_PER_PAGE";
 export const SET_NOTIFICATION_OFFSET = "SET_NOTIFICATION_OFFSET";
@@ -60,28 +60,26 @@ export function updateNotification(showLoading = false, offset) {
     offset
   };
 }
-export function fetchNotifications(showLoading = false, offset, limit) {
+export function fetchNotifications(showLoading = false, notificationPage) {
   return {
     type: FETCH_ALL_NOTIFICATIONS,
     showLoading,
-    offset,
-    limit
+    notificationPage
   };
 }
 
 export function fetchNotificationCounts(showLoading = false) {
   return {
-    type: FETCH_NOTIFICATION_COUNTS,
+    type: FETCH_NOTIFICATION_COUNT,
     showLoading
   };
 }
 
-export function markNotificationAsRead(notificationId, offset, limit) {
+export function markNotificationAsRead(notificationId, notificationPage) {
   return {
     type: MARK_NOTIFICATION_AS_READ,
     notificationId,
-    offset,
-    limit
+    notificationPage
   };
 }
 
@@ -105,26 +103,18 @@ export function fetchHistoryItems(project, offset, limit) {
   };
 }
 
-export function markMultipleNotificationsAsRead(notificationIds, offset, limit) {
+export function markMultipleNotificationsAsRead(notificationIds, notificationPage) {
   return {
-    type: MARK_MULTIPLE_NOTIFICATION_AS_READ,
+    type: MARK_MULTIPLE_NOTIFICATIONS_AS_READ,
     notificationIds,
-    offset,
-    limit
+    notificationPage
   };
 }
 
-export function setNotifcationsPerPage(limit) {
+export function setNotifcationsPerPage(notificationPageSize) {
   return {
     type: SET_NOTIFICATIONS_PER_PAGE,
-    limit
-  };
-}
-
-export function setNotificationOffset(offset) {
-  return {
-    type: SET_NOTIFICATION_OFFSET,
-    offset
+    notificationPageSize: notificationPageSize
   };
 }